refactor round evaluation
